requirements.txt
certifi==2019.3.9
chardet==3.0.4
cloudpayments==1.5.1
Django==2.1
django-cors-headers==2.4.0
django-filter==2.0.0
django-modeladmin-reorder==0.3.1
django-rest-auth==0.9.3
django-tinymce==2.8.0
djangorestframework==3.8.2
idna==2.8
Markdown==2.6.11
pytz==2018.5
requests==2.21.0
six==1.11.0
urllib3==1.24.1
